ollama_rs::generation
Module parameters

Macros

schema_for

Generates a RootSchema for the given type using default settings.

Structs

JsonStructure

Represents a serialized JSON schema. You can create this by converting a JsonSchema:

Enums

FormatType

The format to return a response in

KeepAlive

Used to control how long a model stays loaded in memory, by default models are unloaded after 5 minutes of inactivity

TimeUnit

Traits

JsonSchema

A type which can be described as a JSON Schema document.

Derive Macros

JsonSchema