ollama_rs::generation::tools
Trait Tool

pub trait Tool {
    type Params: Parameters;

    // Required methods
    fn name() -> &'static str;
    fn description() -> &'static str;
    fn call(
        &mut self,
        parameters: Self::Params,
    ) -> impl Future<Output = Result<String, Box<dyn Error + Sync + Send>>>;
}

It’s highly recommended that the JsonSchema has descriptions for all attributes. Descriptions can be defined with #[schemars(descripion = "Hi I am an attribute")] above each attribute

Required Associated Types

type Params: Parameters

Required Methods

fn name() -> &'static str

fn description() -> &'static str

fn call( &mut self, parameters: Self::Params, ) -> impl Future<Output = Result<String, Box<dyn Error + Sync + Send>>>

Call the tool. Note that returning an Err will cause it to be bubbled up. If you want the LLM to handle the error, return that error as a string.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.