ollama_rs::history
Trait ChatHistory

pub trait ChatHistory {
    // Required methods
    fn push(&mut self, message: ChatMessage);
    fn messages(&self) -> Cow<'_, [ChatMessage]>;
}

A trait for managing chat message history.

This trait provides methods for adding messages to the history and retrieving the list of messages.

Required Methods

fn push(&mut self, message: ChatMessage)

Adds a chat message to the history.

Arguments

message - The chat message to be added.

fn messages(&self) -> Cow<'_, [ChatMessage]>

Returns a reference to the list of chat messages in the history.

The messages are returned as a Cow (Clone on Write) to allow for efficient borrowing or cloning as needed.