ollama_rs
Struct Ollama

pub struct Ollama { /* private fields */ }

Implementations

impl Ollama

pub async fn send_chat_messages( &self, request: ChatMessageRequest, ) -> Result<ChatMessageResponse>

Chat message generation. Returns a ChatMessageResponse object

impl Ollama

pub async fn send_chat_messages_with_history<C: ChatHistory>( &mut self, history: &mut C, request: ChatMessageRequest, ) -> Result<ChatMessageResponse>

Chat message generation Returns a ChatMessageResponse object

impl Ollama

pub async fn generate( &self, request: GenerationRequest<'_>, ) -> Result<GenerationResponse>

Completion generation with a single response. Returns a single GenerationResponse object

impl Ollama

pub async fn generate_embeddings( &self, request: GenerateEmbeddingsRequest, ) -> Result<GenerateEmbeddingsResponse>

Generate embeddings from a model

model_name - Name of model to generate embeddings from

prompt - Prompt to generate embeddings for

impl Ollama

pub async fn copy_model( &self, source: String, destination: String, ) -> Result<()>

Copy a model. Creates a model with another name from an existing model.

impl Ollama

pub async fn create_model( &self, request: CreateModelRequest, ) -> Result<CreateModelStatus>

Create a model with a single response, only the final status will be returned.

impl Ollama

pub async fn delete_model(&self, model_name: String) -> Result<()>

Delete a model and its data.

impl Ollama

pub async fn list_local_models(&self) -> Result<Vec<LocalModel>>

impl Ollama

pub async fn pull_model( &self, model_name: String, allow_insecure: bool, ) -> Result<PullModelStatus>

Pull a model with a single response, only the final status will be returned.

model_name - The name of the model to pull.

allow_insecure - Allow insecure connections to the library. Only use this if you are pulling from your own library during development.

impl Ollama

pub async fn push_model( &self, model_name: String, allow_insecure: bool, ) -> Result<PushModelStatus>

Upload a model to a model library. Requires registering for ollama.ai and adding a public key first. Push a model with a single response, only the final status will be returned.

model_name - The name of the model to push in the form of <namespace>/<model>:<tag>.

allow_insecure - Allow insecure connections to the library. Only use this if you are pushing to your library during development.

impl Ollama

pub async fn show_model_info(&self, model_name: String) -> Result<ModelInfo>

Show details about a model including modelfile, template, parameters, license, and system prompt.

impl Ollama

The main struct representing an Ollama client.

This struct is used to interact with the Ollama service.

Fields

url - The base URL of the Ollama service.

reqwest_client - The HTTP client used for requests.

request_headers - Optional headers for requests (enabled with the headers feature).

pub fn new(host: impl IntoUrl, port: u16) -> Self

Creates a new Ollama instance with the specified host and port.

Arguments

host - The host of the Ollama service.

port - The port of the Ollama service.

Returns

A new Ollama instance.

Panics

Panics if the host is not a valid URL or if the URL cannot have a port.

pub fn new_with_client( host: impl IntoUrl, port: u16, reqwest_client: Client, ) -> Self

Creates a new Ollama instance with the specified host, port, and reqwest client.

Arguments

host - The host of the Ollama service.

port - The port of the Ollama service.

reqwest_client - The reqwest client instance.

Returns

A new Ollama instance with the specified reqwest client.

Panics

Panics if the host is not a valid URL or if the URL cannot have a port.

pub fn try_new(url: impl IntoUrl) -> Result<Self, ParseError>

Attempts to create a new Ollama instance from a URL.

Arguments

url - The URL of the Ollama service.

Returns

A Result containing the new Ollama instance or a url::ParseError.

pub fn from_url(url: Url) -> Self

Create new instance from a Url.

pub fn uri(&self) -> String

Returns the URI of the Ollama service as a String.

Panics

Panics if the URL does not have a host.

pub fn url(&self) -> &Url

Returns a reference to the URL of the Ollama service.

pub fn url_str(&self) -> &str

Returns the URL of the Ollama service as a &str.

Syntax in pseudo-BNF:

  url = scheme ":" [ hierarchical | non-hierarchical ] [ "?" query ]? [ "#" fragment ]?
  non-hierarchical = non-hierarchical-path
  non-hierarchical-path = /* Does not start with "/" */
  hierarchical = authority? hierarchical-path
  authority = "//" userinfo? host [ ":" port ]?
  userinfo = username [ ":" password ]? "@"
  hierarchical-path = [ "/" path-segment ]+